Factor out the channels and threads packages
